server/requirements.txt
fastapi==0.115.0
supabase==2.9.0
uvicorn==0.30.0
groq==0.11.0
librosa==0.10.2
scikit-learn==1.5.2
psycopg2-binary==2.9.9
python-telegram-bot==21.5
python-dotenv==1.0.1
gTTS==2.5.3
APScheduler==3.10.4
aiofiles==24.1.0
python-multipart==0.0.12
requests==2.32.3
prophet==1.1.5
haversine==2.8.1
numpy==1.26.4
imaplib2==3.6
Pillow==10.4.0
httpx>=0.27.0,<1.0
